# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-abs",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — the absolute value of `value`.",
        "description": "This service does no arithmetic of its own. It asks `srvcs-isnegative`\nwhether `value` is negative; if so it asks `srvcs-negate` to flip the sign\nand returns that. Otherwise the value is already its own absolute value.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvalResponse"
                }
              }
            }
          },
          "422": {
            "description": "value is not a valid integer (forwarded from a dependency)"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "value"
        ],
        "properties": {
          "value": {
            "type": "object"
          }
        }
      },
      "EvalResponse": {
        "type": "object",
        "required": [
          "value",
          "result"
        ],
        "properties": {
          "result": {
            "type": "object"
          },
          "value": {
            "type": "object"
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      }
    }
  }
}